login

+++ internal/domain/auth/entities/entity.go
@@ -1,10 +1,10 @@
 package entities
 
 type User struct {
-	ID       uint64
-	Mail     string
-	Phone    string
+	ID           int
+	Mail         string
+	Phone        string
 	PasswordHash []byte
-	Role     string
-	Token 	 string
+	Role         string
+	Token        string
 }
